// src/components/Editor/Scenario.action-types.ts
enum ACTION_TYPES {
  CLOSE_EDITOR = "@GIP/CLOSE_EDITOR",
  OPEN_SCENARIO = "@GIP/OPEN_SCENARIO",
  NEW_SCENARIO = "@GIP/NEW_SCENARIO",
  ADD_LINE_AFTER_CURRENT_LINE = "@GIP/ADD_LINE_AFTER_CURRENT_LINE",
  REMOVE_CURRENT_LINE = "@GIP/REMOVE_CURRENT_LINE",
  SET_DIALOG_TITLE = "@GIP/SET_DIALOG_TITLE",
  SET_DIALOG_INTRO = "@GIP/SET_DIALOG_INTRO",
  SWITCH_TO_NEXT_ACTOR = "@GIP/SWITCH_TO_NEXT_ACTOR",
  SWITCH_TO_PREVIOUS_ACTOR = "@GIP/SWITCH_TO_PREVIOUS_ACTOR",
  SWITCH_TO_DESCRIPTION_LINE = "@GIP/SWITCH_TO_DESCRIPTION_LINE",
  SET_FOCUS_ON_NEXT_LINE = "@GIP/SET_FOCUS_ON_NEXT_LINE",
  SET_FOCUS_ON_PREVIOUS_LINE = "@GIP/SET_FOCUS_ON_PREVIOUS_LINE",
  SET_DIALOG_LINE_TEXT = "@GIP/SET_DIALOG_LINE_TEXT",
  SET_SCENARIO_ID = "@GIP/SET_SCENARIO_ID",
  SET_DIRTY = "@GIP/SET_DIRTY",
}
export default ACTION_TYPES;